+++ package.json
@@ -9,6 +9,7 @@
     "@testing-library/user-event": "^12.8.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.1.0"
   },
