+++ src/components/Layout.js
@@ -3,7 +3,6 @@ import { makeStyles } from '@material-ui/core/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import Container from '@material-ui/core/Container';
 import Header from './Header';
-import MainContent from './MainContent';
 import Footer from './Footer';
 
 
@@ -15,24 +14,15 @@ const useStyles = makeStyles((theme) => ({
 
 
 const sections = [
-  { title: 'Файл', url: '#' },
+  { title: 'Файл', url: '/file' },
   { title: 'Моделирование', url: '#' },
   { title: 'Фильтрация', url: '#' },
   { title: 'Анализ', url: '#' },
-  { title: 'Найстройки', url: '#' },
+  { title: 'Настройки', url: '#' },
   { title: 'Справка', url: '#' },
 ];
 
-const mainFeaturedPost = {
-  title:'Проект по компьютерной графике. Лабораторная работа',
-  description:
-    "Визуализация и моделирование многоканальных сигналов",
-  imgText: 'main image description',
-};
-
-
-
-export default function Layout() {
+export default function Layout(props) {
   const classes = useStyles();
 
   return (
@@ -41,7 +31,7 @@ export default function Layout() {
       <Container maxWidth="lg">
         <Header title="CGP - DSP" sections={sections} />
         <main>
-          <MainContent post={mainFeaturedPost} />
+          {props.children}
         </main>
       </Container>
       <Footer description="Еловская И.К., Аликулова З.Х., Идрисов К.И., Ким А.В." />
